Refactor: Change MVI packages
