+++ presentation/src/main/kotlin/io/github/jisungbin/gitmessengerbot/activity/setup/MviSetupState.kt
@@ -2,12 +2,12 @@
  * GitMessengerBot © 2021 지성빈 & 구환. all rights reserved.
  * GitMessengerBot license is under the GPL-3.0.
  *
- * [MviSetupState.kt] created by Ji Sungbin on 21. 9. 19. 오후 4:57
+ * [MviSetupState.kt] created by Ji Sungbin on 21. 9. 19. 오후 5:12
  *
  * Please see: https://github.com/GitMessengerBot/GitMessengerBot-Android/blob/master/LICENSE.
  */
 
-package io.github.jisungbin.gitmessengerbot.activity.setup.mvi
+package io.github.jisungbin.gitmessengerbot.activity.setup
 
 import io.github.jisungbin.gitmessengerbot.mvi.MviBaseState
